[feat] rename toolsAndUtils to component
