+++ app/src/main/java/inu/appcenter/bjj_android/ui/review/component/addImageToList.kt
@@ -1,4 +1,4 @@
-package inu.appcenter.bjj_android.ui.review.toolsAndUtils
+package inu.appcenter.bjj_android.ui.review.component
 
 import android.net.Uri
